refactor: split MovieWriter into focused extensions
